feat: changes the flow for adding an admin

+++ Services/GiftItem/index.js
@@ -65,9 +65,9 @@ const Create = async (data, image) => {
         data: {
           title: data.title,
           details: data.details,
-          amount: parseInt(data.amount),
+          amount: parseFloat(data.amount),
           image: image,
-          weight: parseInt(data.weight),
+          weight: parseFloat(data.weight),
           GiftItemCategory: {
             create: data.categories.map((c) => ({ categoryId: parseInt(c) })),
           },
@@ -100,10 +100,10 @@ const Update = async (id, data, image) => {
       },
       data: {
         image: image ? image : giftItem.image,
-        amount: data.amount ? parseInt(data.amount) : giftItem.amount,
+        amount: data.amount ? parseFloat(data.amount) : giftItem.amount,
         details: data.details ? data.details : giftItem.details,
         title: data.title ? data.title : giftItem.title,
-        weight: data.weight ? parseInt(data.weight) : giftItem.weight,
+        weight: data.weight ? parseFloat(data.weight) : giftItem.weight,
       },
     });
